fix type error for StridedSlice

+++ mindspore/python/mindspore/nn/loss/loss.py
@@ -1464,7 +1464,7 @@ class MultiMarginLoss(LossBase):
         if not weight_one:
             _check_is_tensor('weight', weight, self.cls_name)
         else:
-            weight = self.generate_ones(x.dtype, x[0].shape, 1)
+            weight = self.generate_ones(x.dtype, x.astype('float32')[0].shape, 1)
         loss = self.multi_margin_loss(x, target, weight)
         return loss
